move line routes support trait to a sep file
