+++ moonlight-core/src/main/scala/com/github/jamedge/moonlight/core/api/routes/ApiLineRoutes.scala
@@ -1,47 +1,17 @@
 package com.github.jamedge.moonlight.core.api.routes
 
 import akka.actor.ActorSystem
-import akka.http.scaladsl.marshalling.{Marshaller, ToResponseMarshaller}
 import akka.http.scaladsl.model.StatusCodes
 import akka.http.scaladsl.server.{Directives, Route}
-import akka.http.scaladsl.unmarshalling.FromEntityUnmarshaller
-import akka.stream.Materializer
 import com.github.jamedge.moonlight.core.api.handlers.ApiException
 import com.github.jamedge.moonlight.core.api.versioning.HTMLGenerator
-import com.github.jamedge.moonlight.core.api.versioning.line.{LineHTMLSupport, LineJsonSupport, LineMDGenerator, LineMDSupport}
-import com.github.jamedge.moonlight.core.api.versioning.responsemessage.ResponseMessageJsonSupport
+import com.github.jamedge.moonlight.core.api.versioning.line.{LineMDGenerator, LineRoutesSupport}
 import com.github.jamedge.moonlight.core.model.Line
 import com.github.jamedge.moonlight.core.service.line.LineService
-import scala.concurrent.duration._
 
-import scala.concurrent.{Await, ExecutionContext}
+import scala.concurrent.ExecutionContext
 
 case class ResponseMessage(message: String)
-
-trait LineRoutesSupport {
-  implicit def lineUnmarshaller(implicit matherializer: Materializer): FromEntityUnmarshaller[Line] =
-    LineJsonSupport.unmarshaller
-  implicit def lineMarshaller(
-      implicit ec: ExecutionContext,
-      lineMDGenerator: LineMDGenerator,
-      htmlGenerator: HTMLGenerator
-  ): ToResponseMarshaller[Line] =
-    Marshaller.oneOf(
-      LineJsonSupport.marshaller,
-      LineHTMLSupport.marshaller,
-      LineMDSupport.marshaller
-    )
-  implicit def linesMarshaller(
-      implicit ec: ExecutionContext,
-      lineMDGenerator: LineMDGenerator,
-      htmlGenerator: HTMLGenerator
-  ): ToResponseMarshaller[List[Line]] =
-    Marshaller.oneOf(
-      LineJsonSupport.marshallerLines
-    )
-  implicit def ResponseMessageMarshaller(implicit ec: ExecutionContext): ToResponseMarshaller[ResponseMessage] =
-    ResponseMessageJsonSupport.marshaller
-}
 
 class ApiLineRoutes(
     lineService: LineService
